fix(ConsoleView) : GameStarter에서 view에 해당하는 부분 분리
